% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_MetaModule_score.r
\name{cal_MetaModule}
\alias{cal_MetaModule}
\title{Calculate the MetaModule score for each metacell.}
\usage{
cal_MetaModule(object, module_list, output_path, file_name)
}
\arguments{
\item{object}{A matrix, the row is the gene and the column is the sample/metacell, the value is the expression.}

\item{module_list}{Reaction list or interested gene sets.}

\item{output_path}{Path to save the MetaModule score result.}

\item{file_name}{File name of the MetaModule score result.}
}
\description{
Calculate the MetaModule score for each metacell.
}
