% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{doheatmap_feature}
\alias{doheatmap_feature}
\title{Do the heatmap to show the difference.}
\usage{
doheatmap_feature(object, cluster, feature, width, height)
}
\arguments{
\item{object}{A seurat object.}

\item{cluster}{Info to cluster the cells.}

\item{feature}{Feature to show.}

\item{width}{Width of the figure.}

\item{height}{Height of the figure.}
}
\description{
Do the heatmap to show the difference.
}
